lab6-part2/app/requirements.txt
streamlit 
pypdf2 
langchain
python-dotenv 
faiss-cpu 
openai 
sentencepiece
sentence_transformers
tiktoken
llama-cpp-python